Moved all 'visualise.py' functionality into 'utils.py' for a cleaner file structure
